refactor: Clean up frontend pages - move test pages to alt/ folder
- Move all test and demo pages to pages/alt/ folder
- Keep only 3 main routes: /, /users, /legacy-tool
- Remove unused routes from main.tsx
- Maintain development routes (/login-bypass, /integration-test)

This creates a clean project structure with only production pages visible.
All test pages are safely preserved in alt/ folder for reference.
