+++ frontend/src/pages/alt/CalculatorPage.tsx
